apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]

[project]
name = "kira-api"
version = "0.1.0"
description = "Kira wholesale automation API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "email-validator>=2.2.0,<3.0.0",
  "passlib[argon2]>=1.7.4,<2.0.0",
  "python-jose[cryptography]>=3.3.0,<4.0.0",
  "sqlalchemy>=2.0.35,<3.0.0",
  "psycopg[binary]>=3.2.2,<4.0.0",
  "uvicorn[standard]>=0.31.0,<1.0.0",
  "pydantic-settings>=2.5.2,<3.0.0",
  "openai>=1.47.0,<2.0.0"
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27.2,<1.0.0",
  "pytest>=8.3.3,<9.0.0",
  "ruff>=0.6.9,<1.0.0"
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A", "RUF"]
ignore = ["B008", "E501", "UP017"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
